back_fastapi/requirements.txt
aerich==0.7.1
aiohttp==3.8.3
aiosignal==1.2.0
aiosqlite==0.17.0
anyio==3.6.1
asgi-lifespan==1.0.1
async-generator==1.10
async-timeout==4.0.2
asyncmy==0.2.5
attrs==22.1.0
certifi==2022.9.24
cffi==1.15.1
charset-normalizer==2.1.1
click==8.1.3
colorama==0.4.5
contourpy==1.0.5
cryptography==38.0.1
cycler==0.11.0
DateTime==4.7
Deprecated==1.2.13
dictdiffer==0.9.0
exceptiongroup==1.0.0rc9
fastapi==0.85.1
fonttools==4.38.0
frozenlist==1.3.1
gunicorn==20.1.0
h11==0.14.0
httpcore==0.15.0
httptools==0.5.0
httpx==0.23.0
idna==3.4
importlib-metadata==5.0.0
iniconfig==1.1.1
iso8601==1.1.0
kiwisolver==1.4.4
matplotlib==3.6.1
multidict==6.0.2
numpy==1.23.4
outcome==1.2.0
packaging==21.3
pandas==1.5.1
Pillow==9.2.0
pluggy==1.0.0
py==1.11.0
pycparser==2.21
pydantic==1.10.2
pykrx==1.0.38
pyparsing==3.0.9
pypika-tortoise==0.1.6
pytest==7.1.3
pytest-anyio==0.0.0
pytest-asyncio==0.20.1
pytest-dotenv==0.5.2
pytest-env==0.8.1
pytest-trio==0.7.0
python-dateutil==2.8.2
python-dotenv==0.21.0
pytz==2022.5
PyYAML==6.0
redis==4.3.4
requests==2.28.1
rfc3986==1.5.0
six==1.16.0
sniffio==1.3.0
sortedcontainers==2.4.0
starlette==0.20.4
tomli==2.0.1
tomlkit==0.11.5
tortoise-orm==0.19.2
trio==0.22.0
typing_extensions==4.4.0
urllib3==1.26.12
uvicorn==0.18.3
watchfiles==0.17.0
websockets==10.3
wrapt==1.14.1
xlrd==2.0.1
yarl==1.8.1
zipp==3.9.0
zope.interface==5.5.0
